Fix: Clean up certain comments without names
I have lots of comments that look like this:

-- Name: id; Type: DEFAULT
-- Name: users_pkey; Type: CONSTRAINT

Based on the previous regexes, it looks like the code assumed that at least with certain Postgres versions there's something before the name. But that's not the case for my version (9.5.4). Since I assume this is necessary for some Postgres versions, I've just made the offending part optional.