Allow primary keys to be the last column of a table
Use case: When adding a new column to a table, Postgres adds it to the end of the table definition. If the table has no further constraints, it's impossible to turn this new column into a primary key.

This change fixes the code for the use case so a comma after the primary key column isn't required.