Add the ability to insert deferrable constraints after the table body (#57)
This introduces the `:group_table_definition` option as a more flexible replacement for `:indexes_after_tables`.

It places indexes as well as deferrable constraints after their respective `CREATE TABLE` statements, instead of
later in the file.

The existing option `:indexes_after_tables` is now deprecated and will be removed in a future release.